+++ apps/web/src/components/password-input.tsx
@@ -19,11 +19,12 @@ export function PasswordInput({
   const [showPassword, setShowPassword] = React.useState(false)
 
   return (
-    <div className='relative rounded-md'>
+    <div className='relative rounded-full'>
       <input
         type={showPassword ? 'text' : 'password'}
+        data-slot='input'
         className={cn(
-          'flex h-9 w-full rounded-md border border-input bg-transparent px-3 py-1 pe-9 text-sm shadow-xs transition-colors file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:ring-1 focus-visible:ring-ring focus-visible:outline-hidden disabled:cursor-not-allowed disabled:opacity-50',
+          'flex h-9 w-full rounded-md border border-input bg-transparent px-3 py-1 pe-9 text-sm shadow-xs transition-colors file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:ring-3 focus-visible:ring-ring/50 focus-visible:outline-hidden disabled:cursor-not-allowed disabled:opacity-50',
           className
         )}
         ref={ref}
@@ -35,7 +36,7 @@ export function PasswordInput({
         size='icon'
         variant='ghost'
         disabled={disabled}
-        className='absolute inset-e-1 top-1/2 size-6 -translate-y-1/2 rounded-md text-muted-foreground'
+        className='absolute inset-e-2 top-1/2 size-7 -translate-y-1/2 rounded-full text-muted-foreground'
         onClick={() => setShowPassword((prev) => !prev)}
       >
         {showPassword ? <Eye /> : <EyeOff />}